+++ lib/ts/recipe/oauth2provider/api/utils.ts
@@ -58,6 +58,11 @@ export async function loginGET({
                     },
                     userContext,
                 });
+
+                if ("error" in reject) {
+                    return reject;
+                }
+
                 return { status: "REDIRECT", redirectTo: reject.redirectTo, cookies };
             }
 
@@ -71,6 +76,11 @@ export async function loginGET({
                     },
                     userContext,
                 });
+
+                if ("error" in reject) {
+                    return reject;
+                }
+
                 return { status: "REDIRECT", redirectTo: reject.redirectTo, cookies };
             }
         } catch {
@@ -83,6 +93,11 @@ export async function loginGET({
                 },
                 userContext,
             });
+
+            if (reject.status === "ERROR") {
+                return reject;
+            }
+
             return { status: "REDIRECT", redirectTo: reject.redirectTo, cookies };
         }
     }
@@ -102,6 +117,11 @@ export async function loginGET({
             identityProviderSessionId: session.getHandle(),
             userContext,
         });
+
+        if (accept.status === "ERROR") {
+            return accept;
+        }
+
         return { status: "REDIRECT", redirectTo: accept.redirectTo, cookies: cookies };
     }
 
@@ -126,6 +146,11 @@ export async function loginGET({
             },
             userContext,
         });
+
+        if (reject.status === "ERROR") {
+            return reject;
+        }
+
         return { status: "REDIRECT", redirectTo: reject.redirectTo, cookies };
     }